+++ framework/helpers/ReplaceArrayValue.php
@@ -7,6 +7,8 @@
 
 namespace yii\helpers;
 
+use yii\base\InvalidConfigException;
+
 /**
  * Object that represents the replacement of array value while performing [[ArrayHelper::merge()]].
  *
@@ -69,5 +71,23 @@ class ReplaceArrayValue
     public function __construct($value)
     {
         $this->value = $value;
+    }
+
+    /**
+     * Restores class state after using `var_export()`.
+     *
+     * @param array $state
+     * @return ReplaceArrayValue
+     * @throws InvalidConfigException when $state property does not contain `value` parameter
+     * @see var_export()
+     * @since 2.0.16
+     */
+    public static function __set_state($state)
+    {
+        if (!isset($state['value'])) {
+            throw new InvalidConfigException('Failed to instantiate class "Instance". Required parameter "id" is missing');
+        }
+
+        return new self($state['value']);
     }
 }